Improve zsh plugin check performance for plugins loaded with brew
The brew list command is very slow. Previously, we checked if the oh my
zsh plugins (the ones that have to be installed via brew) are installed
with brew. If a package was installed, we sourced its zsh file, and if
it wasn't, we shown a hint to the user on how to install the plugin.

Now, we try sourcing the plugins' sh file, and if sourcing fails, handle
the error. This is significantly faster and sh startup time improved.

Ask for forgiveness, not permission.